delete long polling logic - add webhook setup
