+++ build.gradle
@@ -2,7 +2,7 @@ plugins {
     id 'java'
 }
 
-group 'com.github.senyast4745'
+group 'com.github'
 version '1.0-SNAPSHOT'
 
 sourceCompatibility = 8
@@ -12,8 +12,9 @@ repositories {
 }
 
 dependencies {
-    compile 'com.sparkjava:spark-core:2.3'
+    compile 'com.sparkjava:spark-core:2.8.0'
     compile group: 'chat.tamtam', name: 'tamtam-bot-api', version: '0.1.9'
+    compile 'org.slf4j:slf4j-simple:1.7.25'
+    compile 'org.slf4j:slf4j-api:1.7.25'
     testCompile group: 'junit', name: 'junit', version: '4.12'
-
 }
